moved diagramState to services
