+++ src/App.js
@@ -1,10 +1,10 @@
-import 'regenerator-runtime/runtime'
-import React, {useState} from 'react'
-import Header from './components/Header'
-import Search from './components/Search'
-import Diagram from './components/Diagram'
+import 'regenerator-runtime/runtime';
+import React from 'react';
+import Header from './components/Header';
+import Search from './components/Search';
+import Diagram from './components/Diagram';
 import ChangeDiagramState from './components/ChangeDiagramState';
-import { DiagramProvider } from './components/DiagramState'
+import { DiagramProvider } from './services/diagramState';
 
 function App() {
   return (
